windows compatibility for killing process

+++ thread_functions.py
@@ -1,7 +1,7 @@
 import bpy , subprocess, threading
 
 from .preferences import get_addon_preferences
-from .misc_functions import absolute_path, create_dir, delete_file, get_file_in_folder
+from .misc_functions import absolute_path, create_dir, delete_file, get_file_in_folder, find_os
 
 # render function
 def render_function(cmd, total_frame, scene, folder_path, output_name, blend_file) :
@@ -13,7 +13,10 @@ def render_function(cmd, total_frame, scene, folder_path, output_name, blend_fil
     frame_count = 0
     while True :
         if not bpy.context.scene.playblaster_is_rendering :
-            process.kill()
+            if find_os() == 'win':
+                subprocess.Popen("TASKKILL /F /PID {pid} /T".format(pid=process.pid))
+            else:
+                process.kill()
             break
         line = process.stdout.readline()
         if line != '' :
